chore : modified function and move enums
