+++ src/components/common/helpers/types.ts
@@ -1,0 +1,13 @@
+export enum FILE_EXTENSION {
+    JSON = 'json',
+    YAML = 'yaml',
+    YML = 'yml',
+}
+
+export enum MIME_TYPE {
+    APPLICATION_JSON = 'application/json',
+    APPLICATION_X_YAML = 'application/x-yaml',
+    APPLICATION_YAML = 'application/yaml',
+    TEXT_X_YAML = 'text/x-yaml',
+    TEXT_YAML = 'text/yaml',
+}
