+++ src/components/common/hooks/types.ts
@@ -26,17 +26,3 @@ export interface FileDataType {
 }
 
 export type ValidatorType = (fileData: FileDataType) => FileReaderStatusType
-
-export enum FILE_EXTENSION {
-    JSON = 'json',
-    YAML = 'yaml',
-    YML = 'yml',
-}
-
-export enum MIME_TYPE {
-    APPLICATION_JSON = 'application/json',
-    APPLICATION_X_YAML = 'application/x-yaml',
-    APPLICATION_YAML = 'application/yaml',
-    TEXT_X_YAML = 'text/x-yaml',
-    TEXT_YAML = 'text/yaml',
-}
